move getLatestSelection to GridSelectionMouseHandler
