+++ packages/iris-grid/src/mousehandlers/IrisGridContextMenuHandler.tsx
@@ -22,8 +22,8 @@ import {
   GridMouseHandler,
   type GridPoint,
   GridRange,
-  type GridRangeIndex,
   GridRenderer,
+  GridSelectionMouseHandler,
   isDeletableGridModel,
   isEditableGridModel,
   isExpandableGridModel,
@@ -148,31 +148,6 @@ class IrisGridContextMenuHandler extends GridMouseHandler {
       cellValue,
       len - command.length - 3
     )}"`;
-  }
-
-  /**
-   * Returns the latest grid selection based on the current grid selection and where the user clicked
-   * This code is dependent on the behavior of GridSelectionMouseHandler.onContextMenu
-   * @param originalSelection The selection from the current grid state which may be stale
-   * @param columnIndex The column index where the user clicked
-   * @param rowIndex The row index where the user clicked
-   */
-  static getLatestSelection(
-    originalSelection: readonly GridRange[],
-    columnIndex: GridRangeIndex,
-    rowIndex: GridRangeIndex
-  ): readonly GridRange[] {
-    const clickedInOriginalSelection = GridRange.containsCell(
-      originalSelection,
-      columnIndex,
-      rowIndex
-    );
-
-    // If the user clicked in a valid cell outside of the original selection,
-    // the selection will be changed to just that cell.
-    return clickedInOriginalSelection || columnIndex == null || rowIndex == null
-      ? originalSelection
-      : [GridRange.makeCell(columnIndex, rowIndex)];
   }
 
   irisGrid: IrisGrid;
@@ -915,7 +890,7 @@ class IrisGridContextMenuHandler extends GridMouseHandler {
       selectedRanges: stateSelectedRanges,
     } = irisGrid.state;
 
-    const selectedRanges = IrisGridContextMenuHandler.getLatestSelection(
+    const selectedRanges = GridSelectionMouseHandler.getLatestSelection(
       stateSelectedRanges,
       columnIndex,
       rowIndex
